refactor: 컴포넌트 index 파일 추가
