+++ src/components/common/Modal/Modal.tsx
@@ -1,0 +1,5 @@
+const Modal: React.FC = () => {
+  return <div>모달</div>;
+};
+
+export default Modal;
