+++ src/components/common/SideBar/SideBar.tsx
@@ -1,0 +1,5 @@
+const SideBar: React.FC = () => {
+  return <div>사이드바</div>;
+};
+
+export default SideBar;
