+++ src/components/index.tsx
@@ -1,0 +1,9 @@
+import Button from '@/components/common/Button/Button';
+import Header from '@/components/common/Header/Header';
+import Image from '@/components/common/Image/Image';
+import Input from '@/components/common/Input/Input';
+import Modal from '@/components/common/Modal/Modal';
+import SideBar from '@/components/common/SideBar/SideBar';
+import Toast from '@/components/common/Toast/Toast';
+
+export { Button, Header, Image, Input, Modal, SideBar, Toast };
