+++ src/pages/home/index.tsx
@@ -1,4 +1,4 @@
-import Header from '@/components/common/Header/Header';
+import { Header } from '@/components/index';
 import PostList from '@/components/home/PostList';
 import { mockPosts } from '@/__mocks__/mockPost';
